Provide content of file in a MemBlock living in VerilogSourceFile.
To be able to have VerilogSourceFile content be used in contexts other
than Analysis (as currently needed with preprocessing), break out the
file content to live in a MemBlock in VerilogSourceFile. That way it
can be used independently.

Then use that all the way down the VerilogAnalyzer chain for the
regular behavior.

Using shared_ptr<MemBlock> for now limits worry about ownership
while refactoring is going on: everyone can have their own (referenced
counted-) copy of the MemBlock, minimizing the changes needed otherwise.
Once the final picture is clear, this might be simplified by having a true
central owner of the MemBlock with guaranteed life-time; then handing
out Stringviews from there is cheaper.

Issues #1502

Signed-off-by: Henner Zeller <hzeller@google.com>