Preserve the sequence of commands provided by the user.
Without that, the commands are just shown alphabetically,
but the user of the subcommands might want to group these
by importance. Always provide 'help' as first command.

While at it: remove the heterogenous lookup ifdefs as we
are using C++17 throughout these days.

Update the "error" text so that calling 'help invalidcommand'
provides a sensible message.

Signed-off-by: Henner Zeller <hzeller@google.com>