Emacs Verilog-Mode AUTOARG expansion in langserver
Adds simple AUTO statement expansion support to the language server.
This feature is modeled after Emacs' Verilog-Mode AUTO expansion. For
now, it is limited to only the AUTOARG statement.

The AUTO expansion is implemented as LSP code actions. Two actions are
currently available: a buffer-wide expansion, which works on the entire
buffer, and a range-based one, which expands all AUTO statement within
the range selected by the user.

Signed-off-by: Krzysztof Bieganski <kbieganski@antmicro.com>