JoinPath(): if path is already absolute, don't prepend anything.
The behavior of JoinPath() used to be that paths are just concatenated
inddpendent if the path in question is already absolute.

  JoinPath("foo/", "bar") -> "foo/bar"  OK
  JoinPath("foo", "/bar") -> "foo/bar"  questionable

This looks like it is a left-over from keeping the behavior the same
from the simplistic old implementation that just concatenated strings.
(https://github.com/chipsalliance/verible/commit/495dda72923d1a13422a6f3a3fd131ed3b1fff37)

However, this results in unexpected behavior. The new behavior is
to keep absolute paths as-is.

  JoinPath("foo", "/bar") -> "/bar"  EXPECTED

Add a test on project tool test that illustrates the problem
that is solved.

Signed-off-by: Henner Zeller <hzeller@google.com>