Make the README in the vscode/ directory more useful.
This README actually shows up in the plugin page in VSCode,
so make it more useful by describing the available features
and where to file bugs.

The original install description, move one directory up to
the general page describing installation for various
editors.

Change version number to 0.0.2700 as we soon hit 2700 commits
after the go-to-definition change is pushed.

Signed-off-by: Henner Zeller <hzeller@google.com>