verilog: tools: ls: symbol-table-handler: Allowed symbol lookup despite errors in symbol table
Since there can be files that have errors in them,
it is not a good idea to stop go-to definition request due to
errors in the symbol table.
This commit removes errors and halting of textDocument/definition request
on Build and Resolve errors.
Instead, it prints warnings and allow functions to continue

Signed-off-by: Grzegorz Latosinski <glatosinski@antmicro.com>