Keep the code clean from using RTTI
There were two instances of using dynamic_cast<> that
could already be implemented with calling Symbol::Kind().

Add a test that compiles without RTTI, so that it is
discovered automatically when a dynamic_cast is added.

For now, keeping the binaries as-is compiled with RTTI,
but they would be about 100kiB smaller without.
Leaving it for another day to check if the performance
improved marginally.

Signed-off-by: Henner Zeller <h.zeller@acm.org>