verilog: analysis: symbol_table: Fixed diagnostics for already existing symbol
SymbolTable during build has a pointer to currently processed source and
string view to symbol.

When symbol is redefined in another file, the source pointer is no longer
valid and resolving the location of the previous symbol in the current
file was resulting in assertion error.

This commit uses the file_origin from SymbolTableNode for previous
symbol to provide symbol location.

Signed-off-by: Grzegorz Latosinski <glatosinski@antmicro.com>