Reduce logging noise: Only aggregate to LOG(INFO), file-level VLOG(1).
For projects with many files, this filled the logs with all kinds
of LOG(WARNING) that are not really relevant unless debugging.
So moved all these to VLOG(1). Keep aggregate logs ("3000 files
took 20 seconds to process") as LOG(INFO).

While at it:
 * Don't call BuildSymbolTable() while in LoadProjectFileList().
   It is confusing as it doesn't do anything, because it requires
   the file to be analyzed first. So just do what is needed to
   open the files into memory.
 * Remove methods only needed internally to the private section
   of SymbolTableHandler header. Also remove unused method.

This makes timings of the parsing more visible, hence touches
feature request #1682

Signed-off-by: Henner Zeller <hzeller@google.com>