Work around an overload that becomes ambiguous with C++20
Starting with c++20, there is a new constructor #5 for
string_view in town:
https://en.cppreference.com/w/cpp/string/basic_string_view/basic_string_view

Together with concepts the initailization with {} can be ambiguous
for either a std::vector<std::string_view> or a std::string_view{}
initialization with two parameters that conceptually allow iterating.

Both would work, in the latter case the last string would be interpreted
as the optional boolean ...

Unfortunately, I could not reproduce this in the CI (#1706 was for that),
I suspect this only shows up with libstdc++ (which we're not using in
the CI).

Long story short: dismbiguiated it by renaming the functions. An alternative
would be to just remove one of the functions.

Signed-off-by: Henner Zeller <hzeller@google.com>