Don't search for .clang-tidy, provide it directly in command line.
If the configuration is not provided in the command line, clang-tidy
will attempt to search for every *.cc file it checks to go up the
directory chain to look for .clang-tidy. That of course might
fail if this is a generated file that lives in a symbolic-linked
bazel-out/ of sorts.

clang-tidy versions 12 and up have a convenient --config-file option,
but clang-tidy 11 only knows --config, so we use that.

While at it, slightly update rules for not wanted diagnosis. Remove
obsolete dtor check variable.

Signed-off-by: Henner Zeller <hzeller@google.com>