Invoke grail compilation db script directly.
Creating the compilation DB from within bazel has
its issues, e.g. we can't include the test targets.

In the project for creating a compilation database for bazel
(https://github.com/grailbio/bazel-compilation-database),
there is also an alternative way: a script that can be used to
create a compilation DB for all targets by invoking bazel
with some aspects enabled.
Let's use that instead: with that, we get all targets, including
test targets into the compilation database.

To avoid having to download the script manually,
hash-sum check  and set up some temporary place on
the users' workstation to host the script and the files it needs, we let
bazel download it in the WORKSPACE file and then
call into it.

While this assumes a certain layout of the bazel
directory structure, this does seem to work fine with
bazel 4, 5, and 6.

Signed-off-by: Henner Zeller <hzeller@google.com>