Remember all files in the clang-tidy cache that were clean.
Previously, we discarded the whole list of files when there
was a single clang-tidy complaint in any of them.
Now, only exclude the files from the cache that had issues,
but keep all the rest. That way it is possible to incrementally
fix all the clang-tidy complatins without having to reprocess
all the files that didn't have any issues.

clang-tidy is very expensive: running it over all verible files
takes about 3-4 core hours. This will help reduce duplicate work.

Signed-off-by: Henner Zeller <hzeller@google.com>