Exclude 'suspicious comma' clang-tidy warning.
This is emitted when there are c-strings concatenated in
sequence, an often occurence in test file. The warning wants to
point out that there are somtimes commas to separate.

Might be useful to address (and some already are), but
given that all these instances are in tests, this is
somewhat lower prio.

Signed-off-by: Henner Zeller <hzeller@google.com>