clang-tidy: content-addressable results in config-dependent dir
Name the directory based on the active configuration, so that
it is possible to play with multiple configurations without
loosing cached information.

For each cc-file, store the output of clang-tidy content-addressed,
so that a minimum of reprocessing is needed: if a file with a
particular content has been processed with a particular clang-tidy,
that content is cached and retrieved quickly.

Make sure that cc files are reprocessed if any of their headers
change.

Provide a symbolic link that points to the last called
clang-tidy output file; thus this can be used as a stable location.

If `CLANG_TIDY` environment variable is set, use its content
as the clang-tidy binary to call, otherwise fall back to search
in system path.

Signed-off-by: Henner Zeller <h.zeller@acm.org>