Show AUTO expansion code actions only for relevant lines
AUTO expansion is now only available if the user invokes the code action
request on a line or range of lines that contains any AUTOs.

Adds a new code action that expands AUTOs only of the same kind as the
ones pointed at/in the selected range.

Adds a separate function that detects if any AUTOs exist in a given
range (`FindAutoKinds()`). This will let us split AUTO detection and
expansion into separate code action and command calls in the future.

Refactors a lot of the code. Now `AutoExpander` only deals with string
views instead of a mix of string views, line ranges, and byte offsets.
In general now the code should be simpler and easier to follow.

Also modifies the test process so that the golden file is formatted
before comparing it to the expanded code.

Fixes #1741.

Signed-off-by: Krzysztof Bieganski <kbieganski@antmicro.com>