Fix c++20 string_view constructor ambiguity.
In C++20, a new constructor for std::string_view (of which absl::string_view is an alias in most cases starting in C++17), causes a double-braced pair of string literals in this context to be treated as a std::vector of a single string_view which is constructed from both literals.

This new constructor is intended to take a pair of begin and end iterators, but can also accept two string literals (since they are iterators of char, after all). In that case, a string_view over all of the intervening memory is returned. That is not the intention here.

Signed-off-by: Henner Zeller <hzeller@google.com>