Rewrite run-clang-tidy script in C++ for readability.
The run-clang-tidy script provides efficient caching
of checked files and their dependencies (as a full run
takes about 400 core minutes: caching is vital).
As such, it is bit more compliated than simply invoking
clang-tidy.

The shell script to deal with that was growing beyond
healthy and was assuming certain features in the
various tools involved (sed, awk, xargs, sort...) that
might not be available everywhere. This, and
complicated quoting rules and esoteric scripting
features also made it fragile, hard to read, and maintain.

This is a rewrite in portable c++17. As such, it is
also much easier to maintain by the C++ authors of
Verible. Tested on Linux and MacOS.

File is designed as C++ script, so can be invoked
directly.

Signed-off-by: Henner Zeller <h.zeller@acm.org>