Canonicalize paths emitted from clang-tidy.
The paths of source files reported are not always relative
to the project root, but some are prefixed with the
bazel root or the current directory.

Canonicalize them so that files are always reported the
same way.

Also: better quote variables in the 'shellscript' compiler invocation
line; update and refine some comments.

Signed-off-by: Henner Zeller <h.zeller@acm.org>