Don't suggest to use emplace_back() all the time.
This is an interesting read: https://abseil.io/tips/112
TL;DR: yes, using emplace_back() can improve performance, but
there can also be surprising effects. So using it might depend
on the context, but it is not necessarily a 'should always'
suggestion (but if we have them in the .clang-tidy, the CI will
fail).

While at it, expand the set of directories that should
be excluded as they don't contain the main project code.

Signed-off-by: Henner Zeller <hzeller@google.com>