Make it more visible if the file list is not found.
If the file list is not found we should show it in the default
logs (VERIBLE_LOGTHRESHOLD=0), not only the verbose logs.
This makes it easier to discover if there are issues (e.g. #1786).

Also remove one default parameter - things are easier to see
when explicitly called.

Now looking at the verilog-language-server_test, and autoexpand_test
it is clear that each of them do some search for a file-list.
In tests, unless we explicitly test for the searching capability,
we always should specify these and not allow auto-searching.
So maybe some infrastructure needed there to provide that at
construction time (even if it is only used in tests).
Left for a later change.

Signed-off-by: Henner Zeller <hzeller@google.com>