Handle incorrect instance types in `AUTOINST`
The `AUTOINST` expansion function assumes the symbol found by searching for the
instance type name is a module declaration. However, if it's not a module
declaration, Verible crashes due to theie assumption. This patch fixes the issue
by checking the type of the node found by the symbol table.

Fixes #1775.

Signed-off-by: Krzysztof Bieganski <kbieganski@antmicro.com>