Change `FirstSubtreeFinder` to extend `SymbolVisitor`
`TreeVisitorRecursive` visits an entire tree, traversing it in pre-order. In
contrast, `SymbolVisitor` only visits a single `Symbol` and needs to manually be
applied for subtrees. This is what already happens in `FirstSubtreeFinder` with
the intention that if it finds the first match, it can return early. However,
`FirstSubtreeFinder` inherits from `TreeVisitorRecursive`, which means it
traverses the entire tree anyway, and the manual applications within it only
create additional, unnecessary work. This patch fixes that by changing
`FirstSubtreeFinder`'s base class to `SymbolVisitor`.

Signed-off-by: Krzysztof Bieganski <kbieganski@antmicro.com>