Prevent expansion of misplaced AUTOs
This patch addresses situations where AUTOs are expanded in incorrect contexts,
such as expanding `AUTOINST` under a port actual, instead of directly in module
instantiation parens.

The syntax tree is traversed until the deepest node that encompasses a given
AUTO expansion is found. If the node kind is expected for the expansion, it can
be applied. Otherwise, the expansion is discarded. For example, `AUTOINST`
expansion can only be done directly under module instantiation, but not any
deeper in the tree, e.g. in a port actual.

Signed-off-by: Krzysztof Bieganski <kbieganski@antmicro.com>