Revise an old filesystem function to return StatusOr<>
When we started, we did not have StatusOr<> available,
but it makes a lot of sense for file::GetContents()
similar to the newer GetContentAsMemBlock().

Provide a new function, similar to it sibling
GetContentAsMemBlock()
  absl::StatusOr<std::string> GetContentAsString();

Provide an adapter function that behaves just as the
old file::GetContents(). To be replaced in subsequent
change(s) until finally removed.

Signed-off-by: Henner Zeller <hzeller@google.com>