clang-tidy: reprocesses troublesome files if older than build env.
If the compilation DB was re-created, it might contain changes that
improve the outcome, e.g. an include directory reference that didn't
work before and made clang-tidy fail on a particular *.cc file
(typically due to BUILD) changes.

Similarly if there are changes in the WORKSPACE file since last
time a file was looked at.

Check timestamps of compile_commands.json and WORKSPACE to
decide if a file that otherwise is unchanged but generated
messges last time (i.e. file size != 0) shall be reprocessed.

Signed-off-by: Henner Zeller <hzeller@google.com>