Support multiple connections to an AUTO-generated port
This patch makes it possible to connect a single module port to multiple
instances during AUTO expansion. For generated ports, this is not a full
solution, as the port's dimensions are taken from the first connected instance.
Instead, it should be the maximum of dimensions of all ports connected to it.

Signed-off-by: Krzysztof Bieganski <kbieganski@antmicro.com>