smoke-test-analyzer: Pin slang to v3.0
Some changes in cmake connfiguration that are not settled yet
make it not work with our script.

Prepare for later slang versions to be compiled with c++20, but
at least for now, pin to version 3.0 to have a stable version
that works in the script and provides a stable comparison.

Signed-off-by: Henner Zeller <hzeller@google.com>