Match `AUTO_TEMPLATE` by instance name regex
The instance name regex in `AUTO_TEMPLATE` has been ignored up until this patch.
Now it is actually being used and considered for matching the correct template
to a given instance. The way it works is straightforward: if a template matches
an instance by type, the instance's name is checked against the template's
regex. If it matches, the template is selected. Otherwise, the next template
that matches by type is considered.

Signed-off-by: Krzysztof Bieganski <kbieganski@antmicro.com>