declaration.cc: verilog.y: Added reshaping for extended function calls
Added reshaping for a case where a function call treated as a data
declaration was not accepting call extensions beyond the first function
call. Adding that change disambiguated the syntax, making it possible to
reshape that call into a proper function call tree structure.
Implemented a method for doing exactly that.

Signed-off-by: Jan Bylicki <jbylicki@antmicro.com>