Nudge pointer alignment towards C/C++ semantic.
This changes the .clang-format to not derive the pointer
alignment from the prevalence in the file, but always make
sure the formatting matches C/C++ semantic (right-aligned).

The files in the code-base used different styles mostly
based on who edited the files first.

The CI clang-format will not enforce it at this point, just
whenever files are edited and cared to be formatted locally,
they get the intended formatting.

Signed-off-by: Henner Zeller <hzeller@google.com>