Break tree_operations test into shards.
This test puts a lot of burden on the compiler and compilation
alone takes about 120 seconds or so. Break it up into shards in
which only parts of the tests are ifdef'ed in in each shard.
Now each shard takes around 20-25 seconds, less likely to dominate
the critical compilation path.

Also, exclude some very slow to clang-tidy tests from run in
clang-tidy.