Less leakage of container type in concrete syntax tree.
Reduces the leaky abstraction of handing out children() by
SyntaxTreeNode, but provide operations to work on them:
front(), back(), operator[].

This change does not get rid of handing out children() entirely yet,
as they are used in range loops, this will be a separate change.

Reducing the abstraction leak will make it easier to change that
implementation later (e.g. for issue #1523)