Building on old platforms gets hard. Just relase the static builds.
Building on a bunch of different platforms often creates trouble,
and with a recent required switch to bazel5 (because of a newer
protobuf version), the java requirements are not met anymore.

However, we already build static binaries that are available for
Linux x86 and arm64 (which are only slightly larger than the
dynamically  linked ones) so let's simplify the release process and just
provide the static binaries.

This makes it simple for everyone who wants to download and run
Verible; for everyone else, they still can compile it themselve
or use the version from their Linux distribution.

Fixes #1871