LSP parse buffer: keep old parses alive whlie calling change listener.
When updating the parse buffers, make sure the old ones don't disappear
while we still call the callbacks for everyone interested in the changes.

They might've hold on to some data and expect the old one to be still
viable when they get the update call.

Add a unit test.

Issues: #2078
While at it: update and refine documentation.