Unregister string view ranges on removed files.
When files are updated and old ones removed, make sure that the
corresponding index for the substring ranges is also updated.

Move the data and logic needed for the substring index into its
own class ContentToFileIndex. Since we _can_ removed string ranges
now, we don't really have to choose anymore if we want to populate
string view maps; still providing this feature for now, but calling
it provide_lookup_file_origin, as this is what it provides.

Make ParsedVerilogSourceFile take an Analyzer instead of a
TextStructureView to better emphasize that this is essentially the same
as the other VerilogSourceFiles, just with a non-owned Analyzer.
This also allows to return the parse absl::Status instead of a
potentially non-truthful ok().

Updating plumbing needed for the passing-an-analyzer.

Add test for updating the project and verify that no stale
string view index is kept (the original issue in #2078).

Fixes: #2078