Make OpenInclude debug messages less noisy.
Only print entering and attempts to open a file as VLOG(2), and
report non-found as VLOG(1). That way it is easier to spot if an include
was not found.

In not-found status, don't record all include paths searched (this
can be a lot, thus a long string), just the number of paths searched.

Refine error reporting in file_util. Turns out that Windows returns
ESRCH ('no such process') when it doesn't find a file.