Implement textDocument/hover request.
Rebased from PR #1922

This will implement the method, but currently it is not offered
yet in the initialization to the editor; to be enabled once tests
finished. The implementation requires a global symbol table, which
is potentially a bit heavy. We might need to introduce the concept of
a per-file symbol table for faster local lookups.