Change name of our json-package at the right place.
I didnn't realize that the name of our json package is not coming
from the nlohmann json source tar, but from our own Bazel file
we made ourself. So changing the name is much simpler there than
providing a patch :)

Remove the patch from #2120 again and just change the name there.

For the work in #2121 though this means, that it would probably better
to use our bazel file (as it uses the single-header version as well
as does proper settings for exceptions)