Don't wrap json generation in jcxxgen() macro, use genrule directly.
This was just to avoid typing a genrule() and a cc_library()
in sequence, while doing that directly would not be much more
work.

The style-guide specifically discourages using macros for sake
of avoiding repetition; From https://bazel.build/rules/bzl-style
  "Do not create a variable or macro just to avoid some amount
   of repetition in BUILD files. Your BUILD file should be easily
   readable both by developers and tools."

Also, move this to common/lsp, as this is the only user.