Fix assumption that string_view iterators are const char *.
Often we construct a string_view with iterator and length.
There is no such constructor for std::string_view that takes an start
iterator and a length. It happens to work in Linux libc++ and absl,
because the string_view iterator is just a const char* (and there is
a constructor std::string_view(const char*, size_t).

Replace all the places that make this assumption and replace with
an existing constructor call.

Also change other places with typename iterator == const char *.