Avoid comparison against nullptr, instead use it as boolean identity.
Readability improvement.

We never use pointers as an artifact of a low-level memory implementation,
we always use it in the context of things that can exist or not.

For these, comparing against nullptr uses a low-level technical aspect
about the implementation, but not what we _mean_.

Using the boolean evaluation is more expressive as it reads more
naturally like an existence check `if (thing) thing->dosomething()`

C/C++ gives us the improved expressiveness, let's use it.