Use single-header json implementation.
clang-tidy complains if a symbol is not in a header directly included
with the [misc-include-cleaner] check. Since the multi-header does
one indirection, this results in these warnings.

Using the single-header library helps. Patch upostream
bzlmod package to provide that.

TODO: check with bzlmod mintainer for nlohmann_json to maybe
provide a target for that.