Use feature of bazel to disable warnings in 3rd party code.
These warnings are not actionable, so this will output more relevant
warnings.

Unfortunately, bazel5 does not understand that yet, it is only a
feature in bazel6 and above.

Does not fix yet the issue in question as we can't enable it quite
yet in the toplevel .bazelrc, so only referencing
Issue #747