Give newline token symbolic name; print raw tokens c-escaped.
The newline token can mean `'\n'`, but also `'\r\n'`, so printing the
token symbolic name as simply `<<\n>>` can be confusing in that context; instead,
print it as symbolic name `<<newline>>`.

Also: the actual raw text of the token is printed in a c-escaped way now,
which makes it easier to human-read if special characters are involved
(and possibly easier to process when grepping through results).