Use new @jsonhpp//:singleheader-json also available in 3.11.3.bcr.1
We used to patch the module to allow the single header version of
the library to be available. Not needed anymore thanks to
@lromor who added it in BCR in
https://github.com/bazelbuild/bazel-central-registry/pull/2865

Now, use the same name for the old-WORKSPACE based build and the
new MODULE based build; for that the name we provide in
bazel/jsonhpp.BUILD, and all uses in the repo, had to be changed.

While at it: fix formatting in MODULE.bazel.