Only switch to absl::string_view implementation on Windows.
On gcc and clang-based systems, std::string_view has the iterator
behave as expected by a lot of Verible code, being a const char*.

On windows c++ lib, the iterator is its own object, which does
not decay to const char*, so in that case we need to use the absl::
implementation which behaves like that.

We used to always change to the absl type, but in reality, we only
need to do that on Windows; adapt the abseil patch.

Longer term, we should remove the expectation of the iterator type,
but that requires access to a Windows machine to locally go through
a compile/fix loop.