Don't assume string_view::iterator is const char*
There were a few assumptions in the code that these types are equivalent,
but they are not on all platforms.

The fixes sometimes involve ugly `&*some_iterator` constructs to get to
the underlying location. Some of these should be re-considered after
switching to c++20.

We also have a `string_view_null_iterator()` function to generate
an iterator that is not initialized to be used in place where we
previously used comparison against nullptr.

Issues: #2323