Read version from MODULE.bazel as fallback.
Uses a starlark native.module_version() feature to
extract the version from module if git version is not defined
or the workspace status could not run otherwise (e.g. when
Verible is not a root module).

While at it, replace build-version.py with a shell-script
that does not require Python, making it easier on the system
requirements and also easier to read and maintain.