before-submit.sh: unconditionally update compilation-db
The compilation-db is not only useful for the clang-tidy run, but
also as preparation for `bant dwyu` to see all external dependencies.
Also - it is useful for development purposes of course.

While at it: in get-bant-path.sh: if there is a list of binaries,
make sure to get the last one as that is likely the newer one
(e.g. in cases MODULE.bazel has been updated, but not `bazel clean`
has been issued, the old versioned one will still be there with
bazel6)