verilog: tools: formatter: output formatted files exactly as produced by verible
Newlines `\n` are transformed into `\r\n` by default when running in
Windows systems. This is NOT what we want. For more information see

  13e0cd0c7f35 (If non-mmap: use stdio reading of files instead of std::stream, 2024-10-02)

This was already fixed for inplace formatting because the output file is
opened in binary mode, but the non-inplace path was left untouched. Set
stdout to binary mode too to make the behaviour consistent.