Disable `--//bazel:use_local_flex_bison` that won't work with bazel 8
This works around #2435 : bazel 8 does not allow a simple toolchain
switch with select(). If somone wants to help with bazel knowledge
to re-enable this, undo this patch and send a fix on top.

Having said that, the original reason for this flag used to be
that rules bison and flex didn't compile on Mac and Windows. These
reasons don't seem to apply anymore.

Fixes: #2435